+++ Suporting/csvReader.py
@@ -1,6 +1,6 @@
 from ..Utility.Debug import *
-from ..Utility.Debug import *
-#from GabesPythonToolBox import Debug
+#from ..Utility.Debug import *
+#from GabesPythonToolBox.Utility.Debug import Debug
 
 def process_line(line: str, seperator_symbol: str = ',', float_symbol: str = '.') -> list:
     Debug.log(f"process_line","Header",group="LIB")