+++ exsampleFiles/Debug.py
@@ -1,7 +1,8 @@
 from GabesPythonToolBox.Utility.Debug import *
+from GabesPythonToolBox.Utility.Logger import *
 
 # Enable debugging
-Debug.set_debug_enabled(True)
+Debug.set_debug_enabled(True,verbosity=4)
 
 # Add some custom message groups
 Debug.add_group('LIB', True)
@@ -9,6 +10,7 @@ Debug.add_group('Database', False)
 Debug.add_group('Showcase', True)
 
 Debug.log("\n\n\n", message_type="None", group="Showcase")
+Debug.log("------------------------------------------", message_type="Info", group="Showcase")
  
 # Log different types of messages with different Types
 Debug.log("Custom Header Message", message_type="Header")
@@ -22,6 +24,7 @@ Debug.log("This is an None Format message", message_type="None")
 Debug.log("This is an End message", message_type="End")
 Debug.log("This has a wrong type", message_type="-")
 
+Debug.log("------------------------------------------", message_type="Info", group="Showcase")
 Debug.log("\n\n\n", message_type="None", group="Showcase")
 
 # Using a group filter